Revert "Accept any number of exceptions"
