--- tests/retry_decorator
@@ -1,1 +1,0 @@
-../retry_decorator/
