--- tox.ini
@@ -1,19 +1,0 @@
-[tox]
-distshare={homedir}/.tox/distshare
-envlist=py{26,27,32,33,34,35,36,37}
-skip_missing_interpreters=true
-indexserver=
-    pypi = https://pypi.python.org/simple
-
-[testenv]
-commands=
-    py.test retry_decorator tests --pep8
-deps =
-    py32: pytest<3.0
-    py{26,27,33,34,35,36,37}: pytest>=3.0
-    pytest-pep8==1.0.6
-
-[pytest]
-addopts = -vvl
-pep8maxlinelength=120
-
